Changed Design
The design has been changed to reflect the new objectives given by john.
Traffic lights are not yet implemented, but the main server (now called
system) will accept and process commands, while still being able to send
commands. Code isn't commented yet, but there's not a lot of it so, it
shouldn't be too difficult to work out.